requirements.txt
Django==5.1.4
torch==2.1.0
transformers==4.48.0
librosa==0.10.2.post1
google-generativeai==0.8.3
python-dotenv==1.0.1
moviepy==2.1.2
langchain-groq==0.2.3
mysql-connector-python==9.1.0
